+++ app/src/main/java/com/jin/honey/feature/cart/ui/content/CartContent.kt
@@ -30,7 +30,7 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import coil.compose.AsyncImage
 import com.jin.honey.R
-import com.jin.domain.model.cart.Cart
+import com.jin.domain.cart.model.Cart
 import com.jin.ui.theme.PointColor
 
 @Composable
