+++ app/src/main/java/com/jin/honey/feature/cart/ui/content/CartOptionModifyBottomSheet.kt
@@ -44,9 +44,9 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.jin.honey.R
-import com.jin.domain.model.cart.Cart
-import com.jin.domain.model.cart.IngredientCart
-import com.jin.model2.cart.CartKey
+import com.jin.domain.cart.model.Cart
+import com.jin.domain.cart.model.IngredientCart
+import com.jin.domain.cart.model.CartKey
 import com.jin.ui.theme.PointColor
 
 @OptIn(ExperimentalMaterial3Api::class)
