+++ app/src/main/java/com/jin/honey/feature/category/ui/CategoryScreen.kt
@@ -48,7 +48,7 @@ import com.jin.honey.R
 import com.jin.honey.feature.home.ui.content.headercontent.LocationSearchBottomSheet
 import com.jin.domain.address.model.Address
 import com.jin.domain.address.model.SearchAddress
-import com.jin.domain.model.cart.Cart
+import com.jin.domain.cart.model.Cart
 import com.jin.domain.model.food.Food
 import kotlinx.coroutines.launch
