+++ app/src/main/java/com/jin/honey/feature/ingredient/ui/IngredientViewModel.kt
@@ -8,7 +8,7 @@ import com.jin.domain.usecase.AddIngredientToCartUseCase
 import com.jin.domain.repositories.PreferencesRepository
 import com.jin.domain.usecase.GetIngredientUseCase
 import com.jin.domain.usecase.GetReviewUseCase
-import com.jin.domain.model.cart.Cart
+import com.jin.domain.cart.model.Cart
 import com.jin.domain.model.review.Review
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.MutableStateFlow
