+++ app/src/main/java/com/jin/honey/feature/ingredient/ui/content/IngredientAddedCart.kt
@@ -21,7 +21,7 @@ import androidx.compose.ui.text.buildAnnotatedString
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.withStyle
 import androidx.compose.ui.unit.dp
-import com.jin.domain.model.cart.Cart
+import com.jin.domain.cart.model.Cart
 import com.jin.ui.theme.AddRecipeRippleColor
 import com.jin.ui.theme.PointColor
