+++ app/src/main/java/com/jin/honey/feature/orderdetail/ui/content/OrderDetailCartItems.kt
@@ -32,7 +32,7 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import coil.compose.AsyncImage
 import com.jin.honey.R
-import com.jin.domain.model.cart.Cart
+import com.jin.domain.cart.model.Cart
 import com.jin.ui.theme.OrderDetailBoxBorderColor
 import com.jin.ui.theme.OrderDetailBoxDividerColor
 import com.jin.ui.theme.OrderDetailDeleteIconColor
@@ -43,8 +43,8 @@ fun OrderDetailCartItems(
     modifier: Modifier,
     cartItems: List<Cart>,
     onShowOptionBottomSheet: () -> Unit,
-    onDeleteMenu: (cartItem:  Cart) -> Unit,
-    onDeleteIngredient: (cartItem:  Cart, ingredientName: String) -> Unit
+    onDeleteMenu: (cartItem: Cart) -> Unit,
+    onDeleteIngredient: (cartItem: Cart, ingredientName: String) -> Unit
 ) {
     val interactionSource = remember { MutableInteractionSource() }
