+++ app/src/main/java/com/jin/honey/feature/paymentdetail/ui/content/PayDetailOrderContent.kt
@@ -20,7 +20,7 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.jin.honey.R
-import com.jin.domain.model.cart.Cart
+import com.jin.domain.cart.model.Cart
 import java.text.NumberFormat
 import java.util.Locale
