+++ core/database/src/main/java/com/jin/database/Converters.kt
@@ -3,8 +3,8 @@ package com.jin.database
 import androidx.room.TypeConverter
 import com.google.gson.reflect.TypeToken
 import com.jin.domain.address.model.Address
-import com.jin.domain.model.cart.Cart
-import com.jin.domain.model.cart.IngredientCart
+import com.jin.domain.cart.model.Cart
+import com.jin.domain.cart.model.IngredientCart
 import com.jin.domain.model.food.Ingredient
 import com.jin.domain.model.recipe.RecipeStep
 import com.jin.network.NetworkProvider
