+++ data/src/main/java/com/jin/data/cart/CartRepositoryImpl.kt
@@ -2,10 +2,10 @@ package com.jin.data.cart
 
 import com.jin.database.datasource.CartTrackingDataSource
 import com.jin.database.entities.CartEntity
-import com.jin.domain.repositories.CartRepository
-import com.jin.domain.model.cart.Cart
+import com.jin.domain.cart.CartRepository
+import com.jin.domain.cart.model.Cart
 import com.jin.domain.model.food.CategoryType
-import com.jin.model2.cart.CartKey
+import com.jin.domain.cart.model.CartKey
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
