+++ domain/src/main/java/com/jin/domain/model/order/Order.kt
@@ -1,7 +1,7 @@
 package com.jin.domain.model.order
 
 import com.jin.domain.address.model.Address
-import com.jin.domain.model.cart.Cart
+import com.jin.domain.cart.model.Cart
 import java.time.Instant
 
 data class Order(
