+++ domain/src/main/java/com/jin/domain/model/review/ReviewPreview.kt
@@ -1,6 +1,6 @@
 package com.jin.domain.model.review
 
-import com.jin.domain.model.cart.IngredientCart
+import com.jin.domain.cart.model.IngredientCart
 
 data class ReviewPreview(
     val review: Review,
