+++ domain/src/main/java/com/jin/domain/usecase/ChangeQuantityOfCartUseCase.kt
@@ -1,7 +1,7 @@
 package com.jin.domain.usecase
 
-import com.jin.domain.repositories.CartRepository
-import com.jin.model2.cart.CartKey
+import com.jin.domain.cart.CartRepository
+import com.jin.domain.cart.model.CartKey
 
 class ChangeQuantityOfCartUseCase(private val repository: CartRepository) {
     suspend operator fun invoke(quantityMap: Map<CartKey, Int>): Result<Unit> {
