+++ domain/src/main/java/com/jin/domain/usecase/RemoveIngredientInCartItemUseCase.kt
@@ -1,7 +1,7 @@
 package com.jin.domain.usecase
 
-import com.jin.domain.repositories.CartRepository
-import com.jin.domain.model.cart.Cart
+import com.jin.domain.cart.CartRepository
+import com.jin.domain.cart.model.Cart
 
 class RemoveIngredientInCartItemUseCase(private val repository: CartRepository) {
     suspend operator fun invoke(cartItem: Cart, ingredientName: String) {
